+++ src/MusicManipulations.jl
@@ -16,16 +16,17 @@ include("specific_modules/drums.jl")
 
 include("motifs/notes_with_motifs.jl")
 
-include("staff_printing/musescore.jl")
+include("visuals/musescore.jl")
 import .MuseScore: musescore;
 
 export MuseScore, musescore
 
-# using Requires
-# @require PyPlot="d330b81b-6aea-500a-939a-2ce795aea3ee" begin
-#     include("visualize.jl")
-# end
-
+using Requires
+function __init__()
+    @require PyPlot="d330b81b-6aea-500a-939a-2ce795aea3ee" begin
+        include("visuals/noteplotter.jl")
+    end
+end
 export Jazz, Drums
 
 end